feat: address code review issues in action version hover/quick fix
- Extract shared module (actionVersionUtils.ts) to eliminate duplicated
  parsing, caching, and API logic between hover and code action providers
- Use non-interactive getSession() to prevent auth prompts during
  passive hover/code-action triggers
- Skip dynamic refs containing `${{` expressions to avoid corrupting
  YAML with incorrect replacements
- Suppress Quick Fix for SHA-pinned action refs to preserve supply-chain
  security posture
- Remove unnecessary `md.isTrusted = true` on hover MarkdownString
- Respect CancellationToken in both providers to abort early on stale
  requests
- Fix misleading "latest semver" comment to match actual date-sorted
  tag selection behavior
